% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinytail.R
\docType{package}
\name{shinytail-package}
\alias{shinytail}
\alias{shinytail-package}
\title{shinytail: Tail a File in Shiny}
\description{
Provides a server-side Shiny object for tailing a
file. Makes it easy to display such a file's updates to a Shiny
app in real-time.
}
\author{
\strong{Maintainer}: Cole Arendt \email{cole@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
}

}
