% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tail.R
\name{tail_file}
\alias{tail_file}
\title{Tail a File}
\usage{
tail_file(filename, options = c("-n+1"))
}
\arguments{
\item{filename}{The file that will be tailed}

\item{options}{A character vector of options to be passed to the tail command}
}
\value{
A processx::process object
}
\description{
Beware buffering...
}
